pyproject.toml
[tool.poetry]
name = "adnmusic"
version = "0.1.0"
description = "Musica o ruido mediante secuencias de ADN"
authors = ["Pablo Diaz <pabloandi@gmail.com>"]

[tool.poetry.scripts]
iniciar = "main"

[tool.poetry.dependencies]
python = "^3.6"
biopython = "^1.77"
pygame = "^1.9.6"
mingus = "^0.6.0"
pypubsub = "^4.0.3"
pyopengl = "^3.1.5"
pillow = "^7.2.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
